chore: move helper func into utils
