+++ rlinf/hybrid_engines/weight_syncer/base.py
@@ -27,37 +27,6 @@ SendFn = Callable[[Any], Awaitable[None]]
 RecvFn = Callable[[], Awaitable[Any]]
 
 
-def materialize_tensor(tensor: torch.Tensor | DTensor) -> torch.Tensor:
-    if isinstance(tensor, DTensor):
-        return tensor.full_tensor()
-    assert isinstance(tensor, torch.Tensor), "Expected a torch.Tensor or DTensor"
-    return tensor
-
-
-def normalize_dtype(dtype: torch.dtype | str) -> torch.dtype:
-    if isinstance(dtype, torch.dtype):
-        return dtype
-    if isinstance(dtype, str):
-        mapping = {
-            "float32": torch.float32,
-            "fp32": torch.float32,
-            "float16": torch.float16,
-            "fp16": torch.float16,
-            "bfloat16": torch.bfloat16,
-            "bf16": torch.bfloat16,
-        }
-        key = dtype.lower()
-        if key in mapping:
-            return mapping[key]
-    raise TypeError(f"Unsupported dtype: {dtype}")
-
-
-def normalize_device(device: torch.device | str | None) -> torch.device:
-    if device is None:
-        device = Worker.torch_device_type
-    return device if isinstance(device, torch.device) else torch.device(device)
-
-
 class WeightSyncer(ABC):
     def __init__(self):
         self._sender_initialized: bool = False
